+++ spring-cloud-config-server/src/main/java/org/springframework/cloud/config/server/environment/secretmanager/GoogleSecretManagerAccessStrategy.java
@@ -14,10 +14,21 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.config.server.environment.secretManager;
+package org.springframework.cloud.config.server.environment.secretmanager;
 
-public interface GoogleConfigProvider {
+import java.util.Comparator;
 
-	String getValue(String key, Boolean mandatory);
+import com.google.cloud.secretmanager.v1.Secret;
+import com.google.cloud.secretmanager.v1.SecretVersion;
+
+public interface GoogleSecretManagerAccessStrategy {
+
+	Iterable<Secret> getSecrets();
+
+	String getSecretValue(Secret secret, Comparator<SecretVersion> comparator);
+
+	String getSecretName(Secret secret);
+
+	Boolean checkRemotePermissions();
 
 }
